requirements/base.txt
pytz==2019.1  # https://github.com/stub42/pytz
python-slugify==3.0.2  # https://github.com/un33k/python-slugify
Pillow==6.0.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
redis==3.2.1  # https://github.com/antirez/redis
celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery
mysqlclient==1.4.2.post1
sorl-thumbnail==12.5.0

msgpack-python=0.5.6
python-slugify = 3.0.2


# Django
# ------------------------------------------------------------------------------
django==2.1.7 # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.39.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.3  # https://github.com/django-compressor/django-compressor
django-redis==4.10.0  # https://github.com/niwinz/django-redis
django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat
django-celery-email==2.0.2
django-test-plus==1.2.1
django-taggit== 1.1.0
django-contrib-comments==1.9.1
django-haystack==2.8.1
django-markdownx==2.0.28

elasticsearch=2.4.1


# Django REST Framework
djangorestframework==3.9.4  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
